rename logging to logger
