+++ doc_scraper/views/main.py
@@ -3,7 +3,7 @@ from flask import request, render_template, make_response, Blueprint
 from doc_scraper.extensions import socketio
 from doc_scraper.settings import INFO_NAMESPACE
 from doc_scraper.errors import ApplicationError
-from doc_scraper.logging import LOGGER
+from doc_scraper.logger import LOGGER
 from doc_scraper.slidedeck import SlideDeck
 from doc_scraper.settings import IMAGE_DIR
